+++ src/main/kotlin/dev/mizarc/waystonewarps/interaction/menus/admin/WarpGroupCreateMenu.kt
@@ -15,6 +15,7 @@ import dev.mizarc.waystonewarps.interaction.utils.lore
 import dev.mizarc.waystonewarps.interaction.utils.name
 import org.bukkit.Material
 import org.bukkit.entity.Player
+import org.bukkit.event.inventory.ClickType
 import org.bukkit.inventory.ItemStack
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
@@ -32,6 +33,8 @@ class WarpGroupCreateMenu(
     override fun open() {
         val gui = AnvilGui(localizationProvider.get(player.uniqueId, LocalizationKeys.MENU_WARP_GROUP_CREATE_TITLE))
         gui.setOnTopClick { it.isCancelled = true }
+        gui.setOnBottomClick { guiEvent -> if (guiEvent.click == ClickType.SHIFT_LEFT ||
+            guiEvent.click == ClickType.SHIFT_RIGHT) guiEvent.isCancelled = true }
         gui.setOnNameInputChanged { newName ->
             if (!isConfirming) groupName = newName else isConfirming = false
         }